Set Router
1. dev server에서 모든 요청이 'index.html'로 가도록 수정
2. RestaurantsPage 컴포넌트 생성
