+++ src/services/api.js
@@ -17,5 +17,6 @@ export async function fetchRestaurants({ regionName, categoryId }) {
     + `?region=${regionName}&category=${categoryId}`;
   const response = await fetch(url);
   const data = await response.json();
+  console.log(data)
   return data;
 }
